Reorganised code into packages. Added documentation for all new classes. Fixed bug when marking tasks as done and reloading it in the next session.
