--- src/main/java/Bye.java
@@ -1,16 +1,0 @@
-public class Bye extends Command {
-
-    public Bye(String input) {
-        super(input);
-    }
-    
-    @Override
-    public void execute(TaskList tasks, Storage storage, Ui ui) {
-        ui.bye();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
