+++ src/main/java/ChatBot.java
@@ -1,3 +1,9 @@
+import chatbot.exceptions.ChatBotException;
+import chatbot.parser.Parser;
+import chatbot.storage.Storage;
+import chatbot.task.TaskList;
+import chatbot.ui.Ui;
+import chatbot.commands.Command;
 
 /**
  * A chat bot that can be renamed, and responds to inputs from users
