+++ src/main/java/chatbot/commands/AddEvent.java
@@ -1,12 +1,24 @@
-public class AddToDo extends Command{
+package chatbot.commands;
 
-    public AddToDo(String input) {
+import chatbot.storage.Storage;
+import chatbot.task.TaskList;
+import chatbot.task.TaskType;
+import chatbot.ui.Ui;
+
+/**
+ * Add an event task to the TaskList
+ * 
+ * @author Owen Yeo
+ */
+public class AddEvent extends Command {
+
+    public AddEvent(String input) {
         super(input);
     }
-
+    
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
-        tasks.addTask(input, TaskType.TODO);
+        tasks.addTask(input, TaskType.EVENT);
 
         ui.print(new String[] {
             "What? You ain't finishing it. Added: ",
