+++ src/main/java/chatbot/commands/DisplayList.java
@@ -1,9 +1,25 @@
+package chatbot.commands;
+
+import chatbot.storage.Storage;
+import chatbot.task.TaskList;
+import chatbot.ui.Ui;
+
+/**
+ * Displays the existing list when executed
+ * 
+ * @author Owen Yeo
+ */
 public class DisplayList extends Command{
 
     public DisplayList(String input) {
         super(input);
     }
     
+    /**
+     * {@inheritDoc}
+     * 
+     * Displays the current list when executed.
+     */
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
         String[] taskStrings = new String[100] ;
