+++ src/main/java/chatbot/exceptions/InvalidDescriptionException.java
@@ -1,3 +1,4 @@
+package chatbot.exceptions;
 /**
  * Exception that is thrown when a command is correct but the description
  * is wrong.
