+++ src/main/java/chatbot/exceptions/InvalidIndexException.java
@@ -1,3 +1,4 @@
+package chatbot.exceptions;
 /**
  * Exception thrown when the index given for mark, unmark, or delete is wrong.
  *
