+++ src/main/java/chatbot/exceptions/SaveFileNotFound.java
@@ -1,3 +1,6 @@
+package chatbot.exceptions;
+import chatbot.exceptions.ChatBotException;
+
 public class SaveFileNotFound extends ChatBotException{
 
     public SaveFileNotFound(String e) {
