+++ src/main/java/data/chatBot.txt
@@ -1,0 +1,2 @@
+E | 1 | cat | 1924-09-11 1900 | 2001-09-11 0900
+D | 0 | chicken | 2400-03-30 0000
